+++ asterion/__init__.py
@@ -2,7 +2,7 @@ from asterion.core.app_factory import create_admin
 from asterion.core.config import CoreAdminConfig
 from asterion.registry import AdminRegistry, ModelAdmin
 
-__version__ = "0.1.3"
+__version__ = "0.1.4"
 
 __all__ = [
     "AdminRegistry",